requirements.txt
Django==1.11.2
djangorestframework==3.6.3
flake8==3.3.0
